Auth Service/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Werkzeug==0.15.0
bcrypt
certifi
cffi
chardet
Flask-SQLAlchemy
idna
pycparser
PyJWT
requests
six
SQLAlchemy
urllib3


